agent_py/requirements.txt
openai>=1.14
openai-agents>=0.5.0
openai-agents-mcp>=0.3.0
fastapi>=0.110
uvicorn[standard]>=0.25
python-dotenv>=1.0
